Update README.md
The previous README.md had two issues which this commit fixes. The previous README.md indicated that clients should pass the access token as Basic authentication. Now, the README.md correctly indicates that the access token should be passed as Bearer authentication. Additionally, the "Get devices" section erroneously indicated that the endpoint to query for a list of possible devices was `/paasapi/abstraction/devices`. Now it correctly indicates that this endpoint is `/paasapi/abstraction/device`.